+++ ChangeLog.md
@@ -1,4 +1,6 @@
 # Change Log
 
 - 2019-11-05 (BF) initial version
+- 2019-11-18 (BF) add transmit :: MonadIO m => PAD -> ReportT m String
+